requirements.txt
opencv-python
numpy
torch
torchvision
timm
openvino-dev
albumentations
matplotlib
colorama